Consider testing your PR
- Fixed compiler errors
- Maximum placement length of tracks and turns are now configurable (#4241)
- Adjusted validation guards and network visualisation to function with longer bends
- Track graph no longer shows nodes and edges on f3 unless configured to
- Kinetic blocks no longer show debug information on f3 unless configured to

+++ src/main/java/com/simibubi/create/content/contraptions/components/structureMovement/AbstractContraptionEntity.java
@@ -745,7 +745,7 @@ public abstract class AbstractContraptionEntity extends Entity implements IEntit
 
 	protected abstract float getStalledAngle();
 
-	protected abstract void handleStallInformation(float x, float y, float z, float angle);
+	protected abstract void handleStallInformation(double x, double y, double z, float angle);
 
 	@OnlyIn(Dist.CLIENT)
 	protected void handleBlockChange(BlockPos localPos, BlockState newState) {
